moved the tests near their scripts
