+++ packages/engine/src/avatar/functions/avatarFunctions.test.ts
