+++ packages/engine/src/ikrig/functions/test-data/pose1/ikrig.pose.bones.ts
