+++ packages/engine/src/ikrig/functions/test-data/rig2.data.ts
