+++ packages/engine/src/ikrig/functions/test-data/rig2.pose.bones-after-hip-legs-spine.ts
