+++ packages/engine/src/map/functions/createIntersectionTestTileCircle.test.ts
