+++ packages/server-core/src/media/storageprovider/storageproviderconfig.ts
